add countries routes by name and by id, error in population

+++ api/src/models/Country.js
@@ -32,7 +32,7 @@ module.exports = (sequelize) => {
       type: DataTypes.STRING,
     },
     Population: {
-      type: DataTypes.INTEGER
+      type: DataTypes.BIGINT,
     }
   });
 };
